refactor: lang of all last archives

+++ src/main/java/br/com/lumilivre/api/dto/aluno/AlunoRankingResponse.java
@@ -3,5 +3,5 @@ package br.com.lumilivre.api.dto.aluno;
 public record AlunoRankingResponse(
 		String matricula,
 		String nome,
-		int emprestimosCount) {
+		long emprestimosCount) {
 }
